Added Linux build configuration.
Modified binding.gyp to find the pg_config command so it can
be used to find the appropriate directories for include files
and library files. Different Linux distributions and PostgreSQL
distributions complicate finding the pg_config command, and
don't always put it in the default PATH, so this change helps
compensate by explicitly searching for the command where it is
likely to be installed.